Fixed vm-bootstrap bug where ctx-agent always ran.
On Nimbus at least the metadata server won't 404 or return empty
string when no userdata is provided. It returns at least a newline
which the script interprets as "True" and feeds into ctx-agent.

This is generally harmless but it may have led to a couple of
failures where both vm-bootstrap and cloudinit.d were attempting
to do git fetch/reset simultaneously.